+++ example/pinecone_example.dart
@@ -1,104 +1,48 @@
-// ignore_for_file: unused_local_variable
+// ignore_for_file: unused_local_variable, avoid_print
+
 import 'dart:io';
 import 'package:pinecone/pinecone.dart';
 
 // Create a client instance with an API key
 final client = PineconeClient(
   apiKey: Platform.environment['PINECONE_API_KEY'] ?? '',
-  environment: 'us-west1-gcp-free',
 );
 
 void main() async {
-  // List indexes
-  List<String> indexNames = await client.index.listIndexes();
+  final environment = 'us-west1-gcp-free';
 
-  // List collections
-  List<String> collectionNames = await client.index.listCollections();
+  /// List indexes
+  final indexes = await client.listIndexes(
+    environment: environment,
+  );
+  print(indexes);
 
-  // Create a new index
-  if (indexNames.isEmpty) {
-    final indexNameToCreate = 'my-index';
-    await client.index.createIndex(
-      body: IndexDefinition(
-        name: indexNameToCreate,
-        dimension: 1536,
-      ),
-    );
-
-    // Poll index until it's ready
-    bool creating = true;
-    while (creating) {
-      print("Waiting for index to be created...");
-      await Future.delayed(Duration(seconds: 10));
-      final index = await client.index.describeIndex(
-        indexName: indexNameToCreate,
-      );
-      creating = index.status.state != IndexState.ready;
-    }
+  if (indexes.isEmpty) {
+    return;
   }
 
-  // Get the latest indexes
-  indexNames = await client.index.listIndexes();
-  print(indexNames);
-
-  // Describe an index
-  final indexMeta = await client.index.describeIndex(
-    indexName: indexNames.first,
-  );
-  print(indexMeta.database.toJson());
-  print(indexMeta.status.toJson());
-
-  // Create a collection
-  if (collectionNames.isEmpty) {
-    await client.index.createCollection(
-      body: CollectionDefinition(
-        name: 'my-collection',
-        source: indexNames.first,
-      ),
+  /// Describe index
+  if (indexes.isNotEmpty) {
+    final index = await client.describeIndex(
+      environment: environment,
+      indexName: indexes.first,
     );
+    print(index);
   }
 
-  // Get the latest collections
-  collectionNames = await client.index.listCollections();
-  print(collectionNames);
-
-  // Describe a collection
-  final collection = await client.index.describeCollection(
-    collectionName: collectionNames.first,
+  /// List collections
+  final collections = await client.listCollections(
+    environment: environment,
   );
-  print(collection.toJson());
+  print(collections);
 
-  // Create sample vector data
-  final vectorData =
-      List.generate(indexMeta.database.dimension, (i) => i.toDouble());
-
-  // Add a vectors to an index and namespace
-  final upsertRes = await client.vector(host: indexMeta.status.host).upsert(
-        body: UpsertRequest(
-          vectors: [
-            for (int i = 0; i < 5; i++)
-              UpsertVector(
-                id: DateTime.now().toString(),
-                values: vectorData,
-              ),
-          ],
-          namespace: 'my-namespace',
-        ),
-      );
-  print(upsertRes.toJson());
-
-  // Query the vector database
-  final queryRes = await client.vector(host: indexMeta.status.host).query(
-        body: QueryRequest(
-          topK: 5,
-          namespace: 'my-namespace',
-          vector: vectorData,
-        ),
-      );
-  print(queryRes.toJson());
-
-  if (queryRes.matches.isNotEmpty) {
-    print(queryRes.matches.first.toJson());
+  /// Describe collection
+  if (collections.isNotEmpty) {
+    final collection = await client.describeCollection(
+      environment: environment,
+      collectionName: collections.first,
+    );
+    print(collection);
   }
 
   client.endSession();