+++ oas/schema/upsert_response.dart
@@ -1,0 +1,10 @@
+part of pinecone_schema;
+
+/// Vector query response object
+final schemaUpsertResponse = Schema.object(
+  properties: {
+    'upsertedCount': Schema.integer(
+      description: 'The number of vectors upserted.',
+    ),
+  },
+);