+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0
+
+* Update for Dart 3 support
+* Use `openapi_spec` package to generate schema/client
+* **Breaking** New client interface to match `openapi_spec` generated code
+
 ## 0.2.0
 
 * Update documentation and API spec